+++ apps/mail/types/theme.ts
@@ -1,5 +1,5 @@
 // Theme types for the frontend (matching the server schema)
-
+import { GoogleFont, defaultTheme } from '@/packages/shared/theme-constants';
 export interface ThemeColors {
   primary: string;
   primaryForeground: string;
@@ -137,96 +137,3 @@ export const googleFonts = [
   'Crimson Text',
   'Space Grotesk',
 ] as const;
-
-export type GoogleFont = (typeof googleFonts)[number];
-
-// Default theme configuration
-export const defaultTheme: Theme = {
-  name: 'Default',
-  description: 'Default application theme',
-  isPublic: false,
-  colors: {
-    primary: 'hsl(240 5.9% 10%)',
-    primaryForeground: 'hsl(0 0% 98%)',
-    secondary: 'hsl(240 4.8% 95.9%)',
-    secondaryForeground: 'hsl(240 5.9% 10%)',
-    background: 'hsl(0 0% 100%)',
-    foreground: 'hsl(240 10% 3.9%)',
-    muted: 'hsl(240 4.8% 95.9%)',
-    mutedForeground: 'hsl(240 3.8% 46.1%)',
-    accent: 'hsl(240 4.8% 95.9%)',
-    accentForeground: 'hsl(240 5.9% 10%)',
-    destructive: 'hsl(0 84.2% 60.2%)',
-    destructiveForeground: 'hsl(0 0% 98%)',
-    border: 'hsl(240 5.9% 90%)',
-    input: 'hsl(240 5.9% 90%)',
-    ring: 'hsl(240 10% 3.9%)',
-    card: 'hsl(0 0% 100%)',
-    cardForeground: 'hsl(240 10% 3.9%)',
-    popover: 'hsl(0 0% 100%)',
-    popoverForeground: 'hsl(240 10% 3.9%)',
-  },
-  typography: {
-    fontFamily: 'Inter, system-ui, sans-serif',
-    fontSize: {
-      xs: '0.75rem',
-      sm: '0.875rem',
-      base: '1rem',
-      lg: '1.125rem',
-      xl: '1.25rem',
-      '2xl': '1.5rem',
-      '3xl': '1.875rem',
-      '4xl': '2.25rem',
-    },
-    fontWeight: {
-      thin: '100',
-      light: '300',
-      normal: '400',
-      medium: '500',
-      semibold: '600',
-      bold: '700',
-      extrabold: '800',
-    },
-    lineHeight: {
-      tight: '1.25',
-      normal: '1.5',
-      relaxed: '1.625',
-      loose: '2',
-    },
-  },
-  spacing: {
-    xs: '0.5rem',
-    sm: '0.75rem',
-    md: '1rem',
-    lg: '1.5rem',
-    xl: '2rem',
-    '2xl': '3rem',
-    '3xl': '4rem',
-    '4xl': '6rem',
-  },
-  shadows: {
-    sm: '0 1px 2px 0 rgb(0 0 0 / 0.05)',
-    base: '0 1px 3px 0 rgb(0 0 0 / 0.1), 0 1px 2px -1px rgb(0 0 0 / 0.1)',
-    md: '0 4px 6px -1px rgb(0 0 0 / 0.1), 0 2px 4px -2px rgb(0 0 0 / 0.1)',
-    lg: '0 10px 15px -3px rgb(0 0 0 / 0.1), 0 4px 6px -4px rgb(0 0 0 / 0.1)',
-    xl: '0 20px 25px -5px rgb(0 0 0 / 0.1), 0 8px 10px -6px rgb(0 0 0 / 0.1)',
-    '2xl': '0 25px 50px -12px rgb(0 0 0 / 0.25)',
-    inner: 'inset 0 2px 4px 0 rgb(0 0 0 / 0.05)',
-  },
-  borderRadius: {
-    none: '0',
-    sm: '0.125rem',
-    base: '0.25rem',
-    md: '0.375rem',
-    lg: '0.5rem',
-    xl: '0.75rem',
-    '2xl': '1rem',
-    '3xl': '1.5rem',
-    full: '9999px',
-  },
-  backgrounds: {
-    gradient: '',
-    pattern: '',
-    blur: false,
-  },
-};
