+++ apps/server/src/routes/autumn.ts
@@ -4,7 +4,7 @@ import { env } from 'cloudflare:workers';
 import { Hono } from 'hono';
 
 const sanitizeCustomerBody = (body: any) => {
-  let bodyCopy = { ...body };
+  const bodyCopy = { ...body };
   delete bodyCopy.id;
   delete bodyCopy.name;
   delete bodyCopy.email;
